+++ controllers/index.js
@@ -1,11 +1,7 @@
 var db = require('../models/mongoose.js')
 
 exports.default = function(req,res){
-  //will probably replace with db function
-  var db_result = get_drawers();
-  res.render(index,{title:'Manifesto Machine',content:db_result})
-}
-
-function get_drawers(){
-
+  db.get_drawers(function(err,data){
+    res.render('index',{title:'Manifesto Machine',categories:data})
+  })
 }
