+++ package.json
@@ -5,10 +5,12 @@
   "main": "index.js",
   "dependencies": {
     "bootstrap": "^3.3.7",
+    "chroma-js": "^1.3.4",
     "express": "^4.15.2",
     "jquery": "^3.2.1",
     "line-by-line": "^0.1.5",
     "mongoose": "^4.10.5",
+    "nodemon": "^1.11.0",
     "pug": "^2.0.0-rc.2"
   },
   "devDependencies": {
@@ -18,8 +20,9 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --config webpack.config.js",
-    "start": "node server.js",
-    "first-run":"npm run build & npm run start"
+    "start": "node app.js",
+    "first-run": "npm run build & npm run start",
+    "dev": "nodemon app.js"
   },
   "author": "M-ITI, Ricardo Rodrigues",
   "license": "ISC"
