+++ routes/index.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var controller = require('../controllers/data.js');
+var controller = require('../controllers/index.js');
 
 router.get('/',controller.default);
