+++ views/index.pug
@@ -1,9 +1,9 @@
 doctype html
 head
   title= title
-  link(rel='stylesheet' type='text/css' href='w3.css')
-  link(rel='stylesheet' type='text/css' href='index.css')
-  script(type='text/javascript' src='bundle.js')
+  link(rel='stylesheet' type='text/css' href='css/w3.css')
+  link(rel='stylesheet' type='text/css' href='css/index.css')
+  script(type='text/javascript' src='js/bundle.js')
 body(class='w3-display-container')
   #main(class='')
     h1#title= title
